Add unit test infrastructure
New unit tests should be added in an appropriate class in src/test/ and
use JUnit5.

Gradle can be used to run the unit tests with `./gradlew test`. See
README.md for instructions on setting up a run configuration for this
from within IntelliJ.

Travis will run the unit tests as part of every build.